libft fit-in

+++ Libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: chanhale <chanhale@student.42seoul.kr      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/29 10:41:01 by chanhale          #+#    #+#             */
-/*   Updated: 2022/01/04 01:35:58 by chanhale         ###   ########.fr       */
+/*   Updated: 2022/01/04 16:41:01 by chanhale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,23 +14,23 @@ static char	*removespace(char *s);
 
 int	ft_atoi(char *str)
 {
-	int	counter;
-	int	minuscounter;
-	int	result;
+	long long	counter;
+	long long	minuscounter;
+	long long	result;
 
 	result = 0;
 	counter = 0;
-	minuscounter = 1;
+	minuscounter = (long long) 1;
 	str = removespace(str);
 	if (str[counter] != '\0' && (str[counter] == '-'
 			|| str[counter] == '+'))
 		if (str[counter++] == '-')
-			minuscounter = -1;
+			minuscounter = (long long) -1;
 	while (str[counter] != '\0' && str[counter] >= '0'
 		&& str[counter] <= '9')
 	{
-		result *= 10;
-		result += (str[counter] - '0') * minuscounter;
+		result *= (long long) 10;
+		result += (long long)((str[counter] - '0') * (minuscounter));
 		counter++;
 	}
 	return (result);
